refactor(protocol): split the update-protocol crate into types/validation/io (R9R-005)
Split the 619-line lib.rs into ownership-correct submodules — `types` (wire
shapes), `validation` (fail-closed admission), `io` (filenames + atomic file
IO) — re-exported flat so the public API is unchanged. Tests move with their
code. Each file is now well under the 600-line cap.
